+++ types/source/cpu.d.ts
@@ -1,0 +1,5 @@
+export = CPU
+declare class CPU {
+  get name(): string
+  sample(): void | null
+}