+++ types/source/http.d.ts
@@ -1,0 +1,5 @@
+export = HTTP
+declare class HTTP {
+  get name(): string
+  sample(requestQueueTime: number): void
+}