+++ types/source/jobQueue.d.ts
@@ -1,0 +1,5 @@
+export = JobQueue
+declare class JobQueue {
+  get name(): string
+  sample(): number | Promise<number>
+}